Fix integer size mismatch.
* subversion/tests/libsvn_wc/wc-queries-test.c
  (parse_explanation_item): Use size_t for the result of strlen().


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1341142 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/tests/libsvn_wc/wc-queries-test.c
@@ -290,7 +290,7 @@ parse_explanation_item(struct explanation_item **parsed_item,
       if (token && token[0] == '(' && token[1] != '~')
         {
           /* Undo the tokenization to switch parser rules */
-          int token_len = strlen(token);
+          size_t token_len = strlen(token);
 
           if (token + token_len < tmp_end)
             token[token_len] = ' ';
@@ -307,7 +307,7 @@ parse_explanation_item(struct explanation_item **parsed_item,
       if (token && *token == '(' && token[1] == '~')
         {
           /* Undo the tokenization to switch parser rules */
-          int token_len = strlen(token);
+          size_t token_len = strlen(token);
 
           if (token + token_len < tmp_end)
             token[token_len] = ' ';
